+++ javascript/node/selenium-webdriver/example/google_search.js
@@ -1,3 +1,24 @@
+// Copyright 2013 Selenium committers
+// Copyright 2013 Software Freedom Conservancy
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+//     You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+/**
+ * @fileoverview An example WebDriver script.
+ * Usage: node selenium-webdriver/example/google_search.js selenium_jar
+ * where selenium_jar is the path to the standalone Selenium server jar to use.
+ */
+
 var fs = require('fs');
 
 var webdriver = require('..'),
@@ -19,7 +40,7 @@ server.start();
 
 var driver = new webdriver.Builder().
     usingServer(server.address()).
-    withCapabilities({'browserName': 'chrome'}).
+    withCapabilities({'browserName': 'firefox'}).
     build();
 
 driver.get('http://www.google.com');
@@ -31,7 +52,5 @@ driver.wait(function() {
   });
 }, 1000);
 
-driver.quit().addBoth(function() {
-  // Don't shutdown the server until all actions are complete.
-  server.stop();
-});
+driver.quit();
+server.stop();