+++ javascript/node/selenium-webdriver/CHANGES.md
@@ -1,3 +1,11 @@
+## v2.32.0-dev
+
+* Added the `selenium-webdriver/testing` package, which provides a basic
+    framework for writing tests using Mocha. See
+    `selenium-webdriver/example/google_search_test.js` for usage.
+* FIXED: remote.SeleniumServer#stop now shuts down within the active control
+    flow, allowing scripts to finish. Use #kill to shutdown immediately.
+
 ## v2.31.0
 
 * Added an example script.